UCF: run scales D_led

+++ ucf_sports/dim_14/classification/1-NN/cross-validation-NN-def.hpp
@@ -22,7 +22,7 @@ class cv_classify_NN //KTH Cross validation
     const std::string actionNames;
 
     
-    const int scale_factor;
+    const float scale_factor;
     const int shift;
     const int dim;
     int p;
