+++ ucf_sports/dim_14/classification/1-NN/cross-validation-NN-impl.hpp
@@ -4,7 +4,7 @@ inline
 cv_classify_NN::cv_classify_NN(const std::string in_path,
 			       const std::string in_path_dataset,
 			       const std::string in_actionNames,  
-			       const int in_scale_factor, 
+			       const float in_scale_factor, 
 			       const int in_shift,
 			       const int in_dim 
 ):path(in_path), path_dataset(in_path_dataset), actionNames(in_actionNames), scale_factor(in_scale_factor), shift(in_shift), dim(in_dim)
